Implement nested comments using redis and get calls to find the appropriate items in the list
I did this because I was too lazy to look up different data sets or migrate to using mongodb, I was worried that implementing the 0auth token part would be difficult with mongodb, but I dont think it would have. Anyway this will do for now. Performance on this will get worse as there get to be more comments, but as I said this will do for now, if that ever becomes a problem I will come back to this when required. MVP and all. Anyway that could be a good challenge for me to learn, tokens, authorisation e.t.c.